Update how-to-fetch-pull-request-number-from-github.md
Corrected a small typo. Also, would that be better to put it like "So, new environment variables can be ..." if we are talking about new environment variables?
Additionally, I am a big fan of showing the output of the presented code so that the reader knows what to expect from a script. Is it possible for someone to add the output of the presented Ruby script?